pyproject.toml
[project]
name = "lyric-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "tabulate>=0.9.0",
    "thefuzz>=0.22.1",
]
